+++ sopalin/starpu/codelet_blok_zgemmsp.c
@@ -346,7 +346,8 @@ starpu_task_blok_zgemmsp( sopalin_data_t   *sopalin_data,
         if ( cblk->ownerid == sopalin_data->solvmtx->clustnum ) {
             need_submit = 1;
         }
-        if ( fcblk->ownerid == sopalin_data->solvmtx->clustnum ) {
+        if ( (fcblk->cblktype & CBLK_FANIN) ||
+             (fcblk->ownerid == sopalin_data->solvmtx->clustnum) ) {
             need_submit = 1;
         }
         else {
@@ -359,12 +360,6 @@ starpu_task_blok_zgemmsp( sopalin_data_t   *sopalin_data,
             return;
         }
     }
-#endif
-
-#if !defined(HAVE_STARPU_DATA_PARTITION_CLEAN_NODE)
-    /* Mark the read cblk as partitionned */
-    cblk->partitioned |= (sideA + 1);
-    cblk->partitioned |= (sideB + 1);
 #endif
 
     /*